refactor: move CSVAdditionalClinicalDataRepository class to outer layer infrastructure
